requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-timeout==2.2.0
mysql-connector-python==8.2.0
webdriver-manager==4.0.1
